# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResult:
    properties:
      message:
        example: Error message
        type: string
      status:
        example: Error
        type: string
    type: object
  models.GenericErrorResponse:
    properties:
      result:
        $ref: "#/definitions/models.ErrorResult"
    type: object
...
  models.RepetitiveTaskTemplate:
    properties:
      createdAt:
        type: string
      description:
        type: string
      friday:
        type: boolean
      id:
        type: string
      isActive:
        type: boolean
      lastDateOfTaskGeneration:
        type: string
      modifiedAt:
        type: string
      monday:
        type: boolean
      priority:
        type: integer
      saturday:
        type: boolean
      schedule:
        type: string
      shouldBeScored:
        type: boolean
      space:
        $ref: "#/definitions/models.Space"
      spaceId:
        type: string
      sunday:
        type: boolean
      tags:
        items:
          $ref: "#/definitions/models.Tag"
        type: array
      tasks:
        items:
          $ref: "#/definitions/models.Task"
        type: array
      thursday:
        type: boolean
      timeOfDay:
        type: string
      title:
        type: string
      tuesday:
        type: boolean
      userId:
        description: Add UserID here
        type: string
      wednesday:
        type: boolean
    required:
      - isActive
      - monday
      - priority
      - schedule
      - shouldBeScored
      - title
    type: object
  models.RepetitiveTaskTemplateRequest:
    properties:
      createdAt:
        type: string
      description:
        type: string
      friday:
        type: boolean
      isActive:
        type: boolean
      lastDateOfTaskGeneration:
        type: string
      modifiedAt:
        type: string
      monday:
        type: boolean
      priority:
        type: integer
      saturday:
        type: boolean
      schedule:
        type: string
      shouldBeScored:
        type: boolean
      spaceId:
        type: string
      sunday:
        type: boolean
      tags:
        items:
          $ref: "#/definitions/models.Tag"
        type: array
      tasks:
        items:
          $ref: "#/definitions/models.Task"
        type: array
      thursday:
        type: boolean
      timeOfDay:
        type: string
      title:
        type: string
      tuesday:
        type: boolean
      wednesday:
        type: boolean
    required:
      - createdAt
      - friday
      - isActive
      - modifiedAt
      - monday
      - priority
      - saturday
      - schedule
      - shouldBeScored
      - sunday
      - thursday
      - title
      - tuesday
      - wednesday
    type: object
  models.RepetitiveTaskTemplateResponseForSwagger:
    properties:
      message:
        example: Success message
        type: string
      result:
        $ref: "#/definitions/models.RepetitiveTaskTemplate"
      status:
        example: Success
        type: string
    type: object
...
  models.Space:
    properties:
      createdAt:
        type: string
      id:
        type: string
      modifiedAt:
        type: string
      name:
        type: string
      userId:
        type: string
    required:
      - createdAt
      - modifiedAt
      - name
    type: object
...
  models.Tag:
    properties:
      createdAt:
        type: string
      id:
        type: string
      modifiedAt:
        type: string
      name:
        type: string
      userId:
        type: string
    required:
      - createdAt
      - modifiedAt
      - name
    type: object
...
  models.Task:
    properties:
      completionStatus:
        type: string
      createdAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      modifiedAt:
        type: string
      priority:
        type: integer
      repetitiveTaskTemplate:
        $ref: "#/definitions/models.RepetitiveTaskTemplate"
      repetitiveTaskTemplateId:
        type: string
      schedule:
        type: string
      score:
        type: integer
      shouldBeScored:
        type: boolean
      space:
        $ref: "#/definitions/models.Space"
      spaceId:
        type: string
      tags:
        items:
          $ref: "#/definitions/models.Tag"
        type: array
      timeOfDay:
        type: string
      title:
        type: string
      userId:
        description: Add UserID here
        type: string
    required:
      - completionStatus
      - isActive
      - modifiedAt
      - priority
      - schedule
      - shouldBeScored
      - title
    type: object
  models.:
    properties:
      completionStatus:
        type: string
      createdAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      isActive:
        type: boolean
      modifiedAt:
        type: string
      priority:
        type: integer
      repetitiveTaskTemplateId:
        type: string
      schedule:
        type: string
      score:
        type: integer
      shouldBeScored:
        type: boolean
      spaceId:
        type: string
      tags:
        items:
          $ref: "#/definitions/models.Tag"
        type: array
      timeOfDay:
        type: string
      title:
        type: string
    required:
      - completionStatus
      - createdAt
      - isActive
      - modifiedAt
      - priority
      - schedule
      - shouldBeScored
      - title
    type: object
  models.TaskResponseForSwagger:
    properties:
      message:
        example: Success message
        type: string
      result:
        $ref: "#/definitions/models.Task"
      status:
        example: Success
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Blockstracker API
  title: Blockstracker
  version: "1.0"
paths:
...
  /tasks:
    post:
      consumes:
        - application/json
      description: Create a new task with the given details
      parameters:
        - description: Task details
          in: body
          name: task
          required: true
          schema:
            $ref: "#/definitions/models."
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.TaskResponseForSwagger"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
      summary: Create a new task
      tags:
        - tasks
  /tasks/{id}:
    put:
      consumes:
        - application/json
      description: Update an existing task with the given details
      parameters:
        - description: Task ID
          in: path
          name: id
          required: true
          type: string
        - description: Task details
          in: body
          name: task
          required: true
          schema:
            $ref: "#/definitions/models."
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.TaskResponseForSwagger"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
      summary: Update an existing task
      tags:
        - tasks
  /tasks/repetitive:
    post:
      consumes:
        - application/json
      description: Create a new repetitive task template with the given details
      parameters:
        - description: Repetitive task template details
          in: body
          name: task
          required: true
          schema:
            $ref: "#/definitions/models.RepetitiveTaskTemplateRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.RepetitiveTaskTemplateResponseForSwagger"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
      summary: Create a new repetitive task template
      tags:
        - tasks
  /tasks/repetitive/{id}:
    put:
      consumes:
        - application/json
      description: Update an existing repetitive task template with the given details
      parameters:
        - description: Repetitive Task Template ID
          in: path
          name: id
          required: true
          type: string
        - description: Repetitive task template details
          in: body
          name: task
          required: true
          schema:
            $ref: "#/definitions/models.RepetitiveTaskTemplateRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.RepetitiveTaskTemplateResponseForSwagger"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
      summary: Update an existing repetitive task template
      tags:
        - tasks
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"